Trait coaster_nn::Dropout

pub trait Dropout<F>: NN<F> {
    fn new_dropout_config(
        &self, 
        dropout: f32, 
        seed: u64
    ) -> Result<Self::CDROP, Error>;

    fn dropout(
        &self, 
        x: &SharedTensor<F>, 
        result: &mut SharedTensor<F>, 
        config: &Self::CDROP
    ) -> Result<(), Error>;

    fn dropout_grad(
        &self, 
        x: &SharedTensor<F>, 
        x_diff: &SharedTensor<F>, 
        result: &SharedTensor<F>, 
        result_diff: &mut SharedTensor<F>, 
        config: &Self::CDROP
    ) -> Result<(), Error>;
}

Provides the functionality for a Backend to support Dropout operations.

Required Methods

fn new_dropout_config(
&self,
dropout: f32,
seed: u64
) -> Result<Self::CDROP, Error>

Creates a new DropoutConfig, which needs to be passed to further dropout Operations.

fn dropout(
&self,
x: &SharedTensor<F>,
result: &mut SharedTensor<F>,
config: &Self::CDROP
) -> Result<(), Error>

Computes non-linear down-sampling (max Pooling) over the input Tensor x.

Saves the result to result.

fn dropout_grad(
&self,
x: &SharedTensor<F>,
x_diff: &SharedTensor<F>,
result: &SharedTensor<F>,
result_diff: &mut SharedTensor<F>,
config: &Self::CDROP
) -> Result<(), Error>

Computes non-linear down-sampling (max Pooling) over the input Tensor x.

Saves the result to result.