Trait coaster_nn::LogSoftmax

pub trait LogSoftmax<F>: NN<F> {
    fn log_softmax(
        &self, 
        x: &SharedTensor<F>, 
        result: &mut SharedTensor<F>
    ) -> Result<(), Error>;

    fn log_softmax_grad(
        &self, 
        x: &SharedTensor<F>, 
        x_diff: &SharedTensor<F>, 
        result_diff: &mut SharedTensor<F>
    ) -> Result<(), Error>;
}

Provides the functionality for a Backend to support LogSoftmax operations.

Required methods

fn log_softmax(
&self,
x: &SharedTensor<F>,
result: &mut SharedTensor<F>
) -> Result<(), Error>

Computes a logarithmic softmax over the input Tensor x.

Saves the result to result.

fn log_softmax_grad(
&self,
x: &SharedTensor<F>,
x_diff: &SharedTensor<F>,
result_diff: &mut SharedTensor<F>
) -> Result<(), Error>

Computes the gradient of a logarithmic softmax over the input Tensor x.

Saves the result to result_diff.

Loading content...