Trait coaster_nn::SigmoidPointwise

pub trait SigmoidPointwise<F>: NN<F> {
    fn sigmoid_pointwise(&self, x: &mut SharedTensor<F>) -> Result<(), Error>;

    fn sigmoid_pointwise_grad(
        &self, 
        x: &SharedTensor<F>, 
        x_diff: &mut SharedTensor<F>
    ) -> Result<(), Error>;
}

Provides the functionality for pointwise Sigmoid operations (overwrites the input with the result of the operation).

Required Methods

fn sigmoid_pointwise(&self, x: &mut SharedTensor<F>) -> Result<(), Error>

Computes the Sigmoid function over the input Tensor x.

Saves the result back to x.

For a no-memory managed version see sigmoid_pointwise_plain.

fn sigmoid_pointwise_grad(
&self,
x: &SharedTensor<F>,
x_diff: &mut SharedTensor<F>
) -> Result<(), Error>

Computes the gradient of a Sigmoid function over the input Tensor x.

Saves the result back to x_diff.